+++ Web application/packages/generic-components/HISTORY.md
@@ -1,3 +1,4 @@
 # v0.0.4
 
 - GenericModule is now split up into GenericModule and GenericModuleContainer
+- Eventlist shows latest events on the bottom and now has a stick to bottom function.